requirements.txt
tkcomponents~=4.0.2
python-barcode~=0.15.1
